🎉 Update Neovim configuration
Switch back to Ayu Dark. Add vim-js and vim-jsx-pretty for working with
JavaScript files, and add nvim-surround for useful surrounding vim
surround movements.

Set splitbelow and splitright behaviour so that opening a split sets the
right/bottom split to be the active buffer.

Add a bunch of directories and file types to the default file ignore
patterns in Telescope.